# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CBT API Documentation",
        "description": "API Documentation for Computer Based Test (CBT) System",
        "contact": {
            "email": "admin@cbt.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/admin/kelas": {
            "get": {
                "tags": [
                    "Admin - Kelas"
                ],
                "summary": "Daftar semua kelas",
                "operationId": "c7357c0b05f0d13ca21d82b76a45b05a",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Kelas"
                ],
                "summary": "Tambah kelas baru",
                "operationId": "d4ee1e4c4ae2f5a44889e6d11fce7c56",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_kelas"
                                ],
                                "properties": {
                                    "nama_kelas": {
                                        "type": "string",
                                        "example": "Kelas 10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/kelas/{id}": {
            "put": {
                "tags": [
                    "Admin - Kelas"
                ],
                "summary": "Update nama kelas",
                "operationId": "679d0ab9471cebfeaf65289443ad96ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_kelas"
                                ],
                                "properties": {
                                    "nama_kelas": {
                                        "type": "string",
                                        "example": "Kelas 10-A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Kelas"
                ],
                "summary": "Hapus kelas",
                "operationId": "4b16fb5a35d009f86455f84a2ccf733f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/kelas/{id}/users": {
            "get": {
                "tags": [
                    "Admin - Kelas"
                ],
                "summary": "Daftar siswa dalam kelas",
                "operationId": "904ebbdfabe47d0e190ca09f2e8ac18d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/mapel/{kelas_id}": {
            "get": {
                "tags": [
                    "Admin - Mapel"
                ],
                "summary": "Daftar semua mata pelajaran per kelas",
                "operationId": "13330bffc22027fe4c745631ddb6a047",
                "parameters": [
                    {
                        "name": "kelas_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/mapel": {
            "post": {
                "tags": [
                    "Admin - Mapel"
                ],
                "summary": "Tambah mata pelajaran baru",
                "operationId": "b8e6270f20bc349a00296129d4e7940e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_mapel",
                                    "kelas_id"
                                ],
                                "properties": {
                                    "nama_mapel": {
                                        "type": "string",
                                        "example": "Matematika"
                                    },
                                    "kelas_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/mapel/{id}": {
            "put": {
                "tags": [
                    "Admin - Mapel"
                ],
                "summary": "Update mata pelajaran",
                "operationId": "90d3885fdcad085ff953017ea0db3012",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_mapel",
                                    "kelas_id"
                                ],
                                "properties": {
                                    "nama_mapel": {
                                        "type": "string",
                                        "example": "Fisika"
                                    },
                                    "kelas_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Mapel"
                ],
                "summary": "Hapus mata pelajaran",
                "operationId": "e6319172006fff179a2794732f62f4fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}